src/requirements.txt
pillow
escpos
requests
xmltodict
pygame
cairo-svg